addQRCode method integrated into QR_Scan class.
