+++ .idea/misc.xml
@@ -1,46 +1,41 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
-  <component name="DesignSurface">
-    <option name="filePathToZoomLevelMap">
-      <map>
-        <entry key="..\:/Users/Dallin Dmytryk/Desktop/Git/QR-Scape/app/src/main/res/layout/activity_personal_stats.xml" value="0.20531526836894215" />
-        <entry key="..\:/Users/Dallin Dmytryk/Desktop/Git/QR-Scape/code/app/src/main/res/drawable/colorgradient.xml" value="0.3828125" />
-        <entry key="..\:/Users/Dallin Dmytryk/Desktop/Git/QR-Scape/code/app/src/main/res/drawable/edit_text_background.xml" value="0.3828125" />
-        <entry key="..\:/Users/Dallin Dmytryk/Desktop/Git/QR-Scape/code/app/src/main/res/layout/activity_login_test.xml" value="0.25" />
-<<<<<<< HEAD
-        <entry key="..\:/Users/Owner/AndroidStudioProjects/QR-Scape/app/src/main/res/layout/activity_personal_stats.xml" value="0.146875" />
-        <entry key="app/src/main/res/layout/qr_collections_layout.xml" value="0.11851851851851852" />
-=======
-        <entry key="..\:/Users/Dallin/Desktop/GIT/QR-Scape/app/src/main/res/layout/activity_home.xml" value="0.1286231884057971" />
-        <entry key="..\:/Users/Dallin/Desktop/GIT/QR-Scape/app/src/main/res/layout/activity_location.xml" value="0.1286231884057971" />
-        <entry key="..\:/Users/Dallin/Desktop/GIT/QR-Scape/app/src/main/res/layout/activity_login.xml" value="0.1286231884057971" />
-        <entry key="..\:/Users/Dallin/Desktop/GIT/QR-Scape/app/src/main/res/layout/activity_main.xml" value="0.1286231884057971" />
-        <entry key="..\:/Users/Dallin/Desktop/GIT/QR-Scape/app/src/main/res/layout/activity_personal_stats.xml" value="0.1286231884057971" />
-        <entry key="..\:/Users/Dallin/Desktop/GIT/QR-Scape/app/src/main/res/layout/activity_scan_view.xml" value="0.1286231884057971" />
-        <entry key="..\:/Users/Dallin/Desktop/GIT/QR-Scape/app/src/main/res/layout/toolbar_title_layout.xml" value="0.1286231884057971" />
-        <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/layout/activity_home.xml" value="0.12454710144927536" />
-        <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/layout/activity_location.xml" value="0.12454710144927536" />
-        <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/layout/activity_login.xml" value="0.14990942028985507" />
-        <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/layout/activity_main.xml" value="0.12454710144927536" />
-        <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/layout/activity_profile.xml" value="0.20380434782608695" />
-        <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/layout/activity_qr_scan.xml" value="0.25" />
-        <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/layout/activity_search.xml" value="0.12454710144927536" />
-        <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/layout/qr_collections_layout.xml" value="0.12454710144927536" />
-        <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/layout/qr_list_content.xml" value="0.12454710144927536" />
-        <entry key="app/src/main/res/layout/qr_collections_layout.xml" value="0.11851851851851852" />
-        <entry key="app/src/main/res/layout/qr_list_content.xml" value="0.27037037037037037" />
->>>>>>> c3f07e295a36556e01beed0af9b457ae1c649ab6
-      </map>
-    </option>
-  </component>
-  <component name="ExternalStorageConfigurationManager" enabled="true" />
-  <component name="JavadocGenerationManager">
-    <option name="OUTPUT_DIRECTORY" value="$PROJECT_DIR$/../../../QR-Scape/doc" />
-  </component>
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" project-jdk-name="11" project-jdk-type="JavaSDK">
-    <output url="file://$PROJECT_DIR$/build/classes" />
-  </component>
-  <component name="ProjectType">
-    <option name="id" value="Android" />
-  </component>
+    <component name="DesignSurface">
+        <option name="filePathToZoomLevelMap">
+            <map>
+                <entry key="..\:/Users/Dallin Dmytryk/Desktop/Git/QR-Scape/app/src/main/res/layout/activity_personal_stats.xml" value="0.20531526836894215" />
+                <entry key="..\:/Users/Dallin Dmytryk/Desktop/Git/QR-Scape/code/app/src/main/res/drawable/colorgradient.xml" value="0.3828125" />
+                <entry key="..\:/Users/Dallin Dmytryk/Desktop/Git/QR-Scape/code/app/src/main/res/drawable/edit_text_background.xml" value="0.3828125" />
+                <entry key="..\:/Users/Dallin Dmytryk/Desktop/Git/QR-Scape/code/app/src/main/res/layout/activity_login_test.xml" value="0.25" />
+                <entry key="..\:/Users/Dallin/Desktop/GIT/QR-Scape/app/src/main/res/layout/activity_home.xml" value="0.1286231884057971" />
+                <entry key="..\:/Users/Dallin/Desktop/GIT/QR-Scape/app/src/main/res/layout/activity_location.xml" value="0.1286231884057971" />
+                <entry key="..\:/Users/Dallin/Desktop/GIT/QR-Scape/app/src/main/res/layout/activity_login.xml" value="0.1286231884057971" />
+                <entry key="..\:/Users/Dallin/Desktop/GIT/QR-Scape/app/src/main/res/layout/activity_main.xml" value="0.1286231884057971" />
+                <entry key="..\:/Users/Dallin/Desktop/GIT/QR-Scape/app/src/main/res/layout/activity_personal_stats.xml" value="0.1286231884057971" />
+                <entry key="..\:/Users/Dallin/Desktop/GIT/QR-Scape/app/src/main/res/layout/activity_scan_view.xml" value="0.1286231884057971" />
+                <entry key="..\:/Users/Dallin/Desktop/GIT/QR-Scape/app/src/main/res/layout/toolbar_title_layout.xml" value="0.1286231884057971" />
+                <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/layout/activity_home.xml" value="0.12454710144927536" />
+                <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/layout/activity_location.xml" value="0.12454710144927536" />
+                <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/layout/activity_login.xml" value="0.14990942028985507" />
+                <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/layout/activity_main.xml" value="0.12454710144927536" />
+                <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/layout/activity_profile.xml" value="0.20380434782608695" />
+                <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/layout/activity_qr_scan.xml" value="0.25" />
+                <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/layout/activity_search.xml" value="0.12454710144927536" />
+                <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/layout/qr_collections_layout.xml" value="0.12454710144927536" />
+                <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/layout/qr_list_content.xml" value="0.12454710144927536" />
+                <entry key="app/src/main/res/layout/qr_collections_layout.xml" value="0.11851851851851852" />
+                <entry key="app/src/main/res/layout/qr_list_content.xml" value="0.27037037037037037" />
+            </map>
+        </option>
+    </component>
+    <component name="ExternalStorageConfigurationManager" enabled="true" />
+    <component name="JavadocGenerationManager">
+        <option name="OUTPUT_DIRECTORY" value="$PROJECT_DIR$/../../../QR-Scape/doc" />
+    </component>
+    <component name="ProjectRootManager" version="2" languageLevel="JDK_11" project-jdk-name="11" project-jdk-type="JavaSDK">
+        <output url="file://$PROJECT_DIR$/build/classes" />
+    </component>
+    <component name="ProjectType">
+        <option name="id" value="Android" />
+    </component>
 </project>
