+++ app/src/main/java/com/example/qr_scape/QRCollectionActivity.java
@@ -98,79 +98,6 @@ public class QRCollectionActivity extends AppCompatActivity {
 
     }
 
-    /**
-     * Add QR codes to database
-     * @param QRText
-     * @param username
-     * @param latitude
-     * @param longitude
-     * @param photo
-     * @author Ty Greve
-     * @version 2
-     */
-    // Add QRCode Method (To be nest in the scanner class)
-    public void addQRCode(String QRText, String username, double latitude, double longitude, Bitmap photo) {
-        // Validate user input
-        if ((QRText.equals(null)) || (username.equals(null))) {
-            Toast.makeText(QRCollectionActivity.this, "Must fill-in both fields", Toast.LENGTH_SHORT).show();
-        }
-
-        // Create QRCode Instance object
-        QRCode qrCode = new QRCode(QRText, username, latitude, longitude, photo);
-
-
-        if (QRText.length() > 0 && username.length() > 0) {
-
-            // Access a Cloud Firestore instance from your Activity
-            FirebaseFirestore db = FirebaseFirestore.getInstance();
-
-            // Create HashMap for QRCodeInstance and put fields from the qrCode object into it
-            HashMap<String, Object> data = new HashMap<>();
-            data.put("Latitude", qrCode.getLatitude());
-            data.put("Longitude", qrCode.getLongitude());
-            data.put("Photo", null);
-            data.put("Score", qrCode.getScore());
-            data.put("Username", username);
-            data.put("RealHash", qrCode.getQRHash());
-
-            // Create HashMap for QRCodes (real/physical) and put fields into it
-            HashMap<String, Object> data1 = new HashMap<>();
-            data1.put("Score", qrCode.getScore());
-
-            // Store to Firestore the QRCodeInstance
-            // Get reference to Firestore collection and Document ID
-            db.collection("QRCodeInstance").document(qrCode.getQRHashSalted())
-                    .set(data) // Set fields in the Firestore database
-                    .addOnSuccessListener(new OnSuccessListener<Void>() {
-                        @Override
-                        public void onSuccess(Void aVoid) {
-                            Log.d(TAG, "Data has been added successfully!");
-                        }
-                    })
-                    .addOnFailureListener(new OnFailureListener() {
-                        @Override
-                        public void onFailure(@NonNull Exception e) {
-                            Log.d(TAG, "Data could not be added!" + e.toString());
-                        }
-                    });
-
-            // Store to Firestore the QRCode (real/physical)
-            db.collection("QRCodes").document(qrCode.getQRHash())
-                    .set(data1) // Set fields in the Firestore database
-                    .addOnSuccessListener(new OnSuccessListener<Void>() {
-                        @Override
-                        public void onSuccess(Void aVoid) {
-                            Log.d(TAG, "Data has been added successfully!");
-                        }
-                    })
-                    .addOnFailureListener(new OnFailureListener() {
-                        @Override
-                        public void onFailure(@NonNull Exception e) {
-                            Log.d(TAG, "Data could not be added!" + e.toString());
-                        }
-                    });
-        }
-    }//end addQRCode
 
     /**
      * Deletes an instance of a user QR code from the database
